requirements.txt
fastapi==0.60.1
cmake
pillow==7.1.2
aiofiles==0.4.0
python-multipart==0.0.5 
aiofiles==0.4.0
torchvision==0.9.1
torch==1.8.1  
torchvision==0.9.1 
dlib==19.22.0
opencv-python==4.5.1.48
numpy==1.19.2